+++ modules/deluge.nix
@@ -1,0 +1,5 @@
+{
+  homeModule = { pkgs, ... }: {
+    home.packages = builtins.attrValues { inherit (pkgs) deluge; };
+  };
+}