+++ overlays/nixos-option.nix
@@ -1,1 +1,0 @@
-self: super: { nixos-option = self.emptyDirectory; }